feat(sdk): add IggyStream builder (#1531)
## Summary

The IggyStream, IggyStreamConsumer, and IggyStreamProducer are three
builder that simplify and streamline the process of getting started with
iggy.

## **Rationale**
Similar to Keras for Tensoflow, the IggyStream builder is a way to
simplify the process of getting started with iggy with just the minimum
amount of code while still allowing powerful customization. For a simple
PoC, about 20-30 lines of Rust should result in a working setup.

## **Context**:
Before the IggyStream builder, it was at times hard and confusing to get
started with iggy. For example, server side heartbeat errors occured
because new users forgot to connect the client.

After the IggyStream builder, it is way easier to get started by calling
one of the constructors that returns a correctly configured client that
is already connected and returns a correctly configured producer,
consumer, or both.

For intermediate users, the new configuration types enable the user to
configure the client, producer, and consumer in a way that is easy to
understand and use and yet allows for customization.

Advanced users may skip the builder and configure the client, producer,
and consumer directly using the low-level API.

## **References**:
Based on work conducted in the Iggy Builder repo:
https://github.com/marvin-hansen/iggy_builder

---------

Signed-off-by: Marvin Hansen <marvin.hansen@gmail.com>
Co-authored-by: spetz <piotr.gankiewicz@gmail.com>