chore(sdk): Initial commit with existing iggy-rs/iggy-cpp-client code (#1775)
As-is migration of iggy-cpp-client. I will bring packages up to newest
versions and do some light clean-up prior to doing the larger Bazel migration.
This SDK was never fully completed, and more recently we have
discussed changing approach and using WASM, so this is really just
getting a clean baseline.

Fixes #1733

---------

Co-authored-by: Hubert Gruszecki <h.gruszecki@gmail.com>