chore(ci): Remove archived actions-rs/* actions (#1858)
The actions-rs org has been archived for years at this point and it
doesn't look like it's coming back. For security reasons we are going to
remove it from the list of allowed actions in the org soon. (See
https://github.com/apache/infrastructure-actions)

In this PR I have replaced the usage of `actions-rs/cargo` and
`actions-rs/toolchain` with bare `rustup` and `cargo` commands.