fix(python): fix proper typing for complex enums (#2039)
This updates `pyo3-stub-gen` and uses the new
`gen_stub_pyclass_complex_enum` to generate proper typing stubs for
complex enums, now the entire test file type checks! (Still some `Any`
where we want proper types but that has to wait for my upstream PR).

I've also taken the liberty to add myself to the authors lists of the
Python SDK and the version has been bumped to prepare for the upcoming
release.