refactor(csharp): refactor method declaration for csharp sdk (#2108)
Currently, contract methods are different then in other sdks. After
changes it's built in the same way. It's remove unnecesery object
declaration for method invokation.

This breaks almost all methods.

What has been done:
- refactored contract methods 
- refactored internal mapping methods
- updates nugets
- apply dotnet format to entire  solution

Also, sending messages and polling was done in very quick way. It will
be changed in one of the next pr. Propably will be created separete
classes for publisher and and consumer with background sending and
polling. In IggyClient will be only raw contracts.

---------

Co-authored-by: Piotr Gankiewicz <piotr.gankiewicz@gmail.com>