chore(example): test python examples in CI (#2174)
* Added one script `scripts/run-python-examples-from-readme.sh` which
runs all python examples and returns an error when any of the examples
fails. For consistency, the script basically aligns with other scripts
for different SDKs.
* Added one github action that builds the python SDK package and runs
example tests. Since checking python examples requires building and
installing the SDK, which means a few more setup steps, the check is
turned into a github action.
* Added one task that checks python examples using the new action above.
* Updated a few things regarding the examples, e.g. fixing the default
argument in the basic example, so that they work with the test script.

---------

Co-authored-by: Hubert Gruszecki <h.gruszecki@gmail.com>