fix(server): fix inconsistent current_position of active segment (#2380)
This PR addresses concurrency issue - when committing journal
`current_position` was in inconsistent state, which lead to incorrect
index generation for pipelined batch.

Also `Murmur3` hash does not deal well with generating hash for small
integer values, leading to skewed `shard_id` assignment for partitions
that are part of the same namespace. Fixed by discarding the lower bits
of the generated hash, when calculating the shard assignment.