feat(cluster): Add deterministic tick-based timeout mechanism for VSR (#2450)
This PR adds `timeout` mechanism to be used in the consensus. 

`Timeout` - is strictly numerical and sync that is it only does counting
down to zero.
`TimeoutManager` - this will be used in `VsrConsensus` and this handles
ticking of all timeouts belonging to a replica.

This tick based timeout system works in two phases phase-1 does
decrement of all `Timeout`s and in 2nd phase we check each `Timeout` and
if it fires we handle by calling respective handler.

We use `xoshiro256+` rng for generating fast random numbers similar to
TB.